+++ app/src/main/java/elite/intel/ai/brain/commons/CommandEndPoint.java
@@ -12,12 +12,10 @@ public abstract class CommandEndPoint extends AiEndPoint {
     private static final Logger log = LogManager.getLogger(CommandEndPoint.class);
     private final AiPromptFactory contextFactory;
     private final AIRouterInterface router;
-    private final AIChatInterface chatInterface;
 
 
     protected CommandEndPoint() {
         this.router = ApiFactory.getInstance().getAiRouter();
-        this.chatInterface = ApiFactory.getInstance().getChatEndpoint();
         this.contextFactory = ApiFactory.getInstance().getAiPromptFactory();
     }
 
@@ -30,6 +28,6 @@ public abstract class CommandEndPoint extends AiEndPoint {
     }
 
     public AIChatInterface getChatInterface() {
-        return chatInterface;
+        return ApiFactory.getInstance().getChatEndpoint();
     }
 }